+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Add scene support. (#118)
 - Add `sceneWhitelist` option for whitelisting scenes. (#118)
+- Add Wireless Switch support (`wxkg`).
+- Add Solar Light support (`tyndj`).
+- Add Dehumidifier support (`cs`).
 
 ### Changed
 - Support Ceiling Fan icon customize and Floor Fan `lock`, `swing` feature. (#131)